+++ src/screens/HomeScreen/_components/forecastsGraph/ForecastsGraphSectionWrapper.type.tsx
